requirements.txt
python-docx==1.1.0
selenium==4.16.0
